prefer glue_data for glue 1.8.0 compatibility
As of v1.8.0, glue::glue aborts if its .envir argument is not an
environment.  yspec passes a list in two spots.  To make these
compatible with v1.8.0, we could 1) convert the list argument to an
environment before passing it to as .envir or 2) call glue_data and
pass the list to its .x argument.  Both those options can express the
same evaluation environments.

In either case, there is a change in behavior here.  With glue
versions before 1.8.0, passing a list to .envir leads to an eval
environment that is undocumented (and likely surprising to the
caller): eval uses the glue:::identity_transformer environment as the
enclosing environment.  The main user-visible consequence is that, if
an item isn't found in the specified list, it can be grabbed from the
global environment.

In the context of yspec, that isn't behavior that should be supported.
The values should be defined by the list coming from the spec.

Convert the incompatible calls to glue_data calls that pass the list
as .x and specify an empty environment.